requirements.txt
six==1.15.0
numpy==1.19.5
protobuf==3.15.8
tensorflow-gpu==2.4.1
tensorboard==2.4.1
nibabel==4.0.2
matplotlib==3.4.0
scikit-image==0.16.2
pandas==1.0.5
SimpleITK==2.0.2
imageio==2.9.0